tests: use github workflow instead of travis
