+++ .github/workflows/tests.yml
@@ -1,0 +1,36 @@
+---
+name: tests
+
+on:
+  pull_request:
+  push:
+  schedule:
+    - cron: '0 6 * * *'
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python: [2.7, 3.6, 3.7, 3.8, 3.9]
+
+    steps:
+
+      - uses: actions/checkout@v2
+
+      - name: Setup Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python }}
+
+      - name: Install Tox and any other packages
+        run: pip install tox
+
+      - name: Run Tox
+        # Run tox using the version of Python in `PATH`
+        run: tox -e py
+
+      # TODO(roaldnefs): Enable after fixing all current syntax issues
+      # - name: Run flake8
+      #   run: tox -e pep8
