--- .travis.yml
@@ -1,24 +1,0 @@
-language: python
-python:
-  - "2.7"
-  - "3.6"
-# command to install dependencie
-install:
-  - python setup.py -q install
-# command to run tests
-script:
-  - python meteoalertapi/test.py
-after_success:
-  - bash <(curl -s https://copilot.blackducksoftware.com/ci/travis/scripts/upload)
-# Deploy a new release to PyPi for every new tag on the 'master' branch on the
-# repository 'rolfberkenbosch/meteoalert-api'
-deploy:
-  provider: pypi
-  user: roaldnefs
-  password:
-    secure: "Mt+ufbsfWrr9HfqzVMAQoUHzGCG6BukklQlQ/vLVN1AiNwIUDbT6yDNfWH+VsoaOq5mpgoEG4FD0OHXlV2pUUAQgTrqcQaQZ4keuteQ4lKmqDmmN5d4o/kmno011aHpGB67Om3pcqmCmH8qJD2WBwc6oNotBZrfrBDdc47hQehe6pTvyz0zgoPXHamHnlbxWbJ21iCcW7fAE0401QKuXF6vRKuHKUbfpkvgUvlBsaEHRzXT/Bu2v4KiuuvHBnhfHBmfv6b1yPlsnI0S8acB/qfUBOPw9mJ0LhQSzHLTr/ljK0YrGQXrzx3Soq8Cs2uSaZN/RNaLPsCuJKxrf0wMBhGohU/ZICpYos0MxocVWp1grf7xebngz6D6vy1kPgfLu5kvyVrWWlmFm/ETCq040kCgMlrzSE3kchJurPFB4hkCc1yohMjEWC9yZGkkV+QTKpmYBWbPndTIKGHS4NtjEu3TUVsvJQcH4n4BMYKvdyDek1wT2ejrN9c33cgtLBBe21eyk2uHI8xPbUYwst0QTEPnOMwG/DWBfdmv5keepkyvpmIZsUx+1tRzV6ao8gFQrnVt6qEqalWT0lZ//rjllUANmrtIsPskOF4ztC922P490GY2WjrrnzyYHjl01yUz17j9BslJQSil1ihu7PfNZoTVM3nd9CxiW/AmCDVg041o="    
-  on:
-    repo: rolfberkenbosch/meteoalert-api
-    tags: true
-    branch: master
-    python: 3.6
