+++ tox.ini
@@ -1,0 +1,44 @@
+[tox]
+envlist = {py27,py36,py37,py38,py39}-install,pep8
+
+[testenv]
+deps =
+  pytest
+  flake8
+commands =
+  pytest tests {posargs}
+
+[testenv:pep8]
+commands =
+  flake8 {posargs} meteoalertapi/ tests/
+
+[testenv:install]
+skip_install = True
+commands =
+    python setup.py bdist_wheel
+    pip install --no-index --find-links=dist meteoalertapi
+
+[testenv:py27-install]
+skip_install = {[testenv:install]skip_install}
+commands = {[testenv:install]commands}
+
+[testenv:py36-install]
+skip_install = {[testenv:install]skip_install}
+commands = {[testenv:install]commands}
+
+[testenv:py37-install]
+skip_install = {[testenv:install]skip_install}
+commands = {[testenv:install]commands}
+
+[testenv:py38-install]
+skip_install = {[testenv:install]skip_install}
+commands = {[testenv:install]commands}
+
+[testenv:py39-install]
+skip_install = {[testenv:install]skip_install}
+commands = {[testenv:install]commands}
+
+[mypy]
+
+[flake8]
+exclude = .git,.venv,.tox,dist,doc,*egg,build,
